phase-2/backend/requirements.txt
fastapi
sqlmodel
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
python-dotenv
bcrypt==4.0.1
openai>=1.0.0
mcp-sdk>=0.0.1
pydantic>=2.0.0

